+++ auntion-product/src/app/admin/chat-admin/chat-admin-home/chat-admin-home.component.ts
@@ -1,0 +1,20 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-chat-admin-home',
+  templateUrl: './chat-admin-home.component.html',
+  styleUrls: ['./chat-admin-home.component.css']
+})
+export class ChatAdminHomeComponent implements OnInit {
+
+  userItemChild: any;
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+  getUserFromChild(event: any) {
+    this.userItemChild = event;
+  }
+}
