+++ auntion-product/src/app/admin/chat-admin/chat-box-admin/chat-box-admin.component.html
@@ -1,0 +1,36 @@
+<div *ngIf="member" class="chat-box__container">
+  <header class="chat-box__header">
+    <div class="info">
+      <p class="name">{{member.nameMember}}</p>
+      <span class="active">
+        <div></div>
+        <span>Active</span>
+      </span>
+    </div>
+
+    <i class="fa-solid fa-ellipsis-vertical"></i>
+  </header>
+
+  <!-- Message of user -->
+  <div class="chat-box__middle">
+    <div class="chat-box__item" *ngFor="let message of messages | async"
+         [ngClass]="message.isOwn.length == 2?'admin':null">
+      <div class="detail">
+        <span class="time">{{message.timeSent}}</span>
+      </div>
+
+      <div class="text-chat">
+        <img class="text-chat__img" *ngIf="message.fileUpload && !message.isFile" [src]="message.fileUpload.url" alt="">
+
+        <a class="text-chat__file" *ngIf="message.fileUpload && message.isFile"
+           href="{{message.fileUpload.url}}" target="_blank">{{message.fileUpload.name}}</a>
+
+        <p>{{message.message}}</p>
+      </div>
+    </div>
+  </div>
+
+  <div class="chat-box__footer">
+    <app-chat-form-admin [userIdChild]="userIdParent"></app-chat-form-admin>
+  </div>
+</div>
