+++ auntion-product/src/app/admin/chat-admin/user-list/user-list.component.css
@@ -1,0 +1,119 @@
+.user-list__container h3 {
+  font-size: 1.5rem;
+  margin: 12px 0;
+  font-weight: 600;
+}
+
+.user-list__container {
+  padding-left: 12px;
+}
+
+.user-list__container .form__search-chat {
+  height: 45px;
+  border-radius: 30px;
+  display: flex;
+  align-items: center;
+  padding: 0 8px;
+  background-color: #ffffff;
+  border: 1px solid #dadada;
+  overflow: hidden;
+}
+
+.user-list__container .form__search-chat input {
+  border: none;
+  outline: none;
+  height: 100%;
+  padding: 0 12px;
+  flex: 2;
+  background-color: #ffffff;
+  display: flex;
+  align-items: center;
+  font-size: 1.1rem;
+}
+
+.user-list__container .form__search-chat i {
+  height: 100%;
+  padding: 0 8px 0px 12px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+  font-size: 1.1rem;
+  font-weight: 600;
+  transition: all 0.3s;
+}
+
+.user-list__container .form__search-chat i:hover {
+  color: var(--button-color);
+}
+
+/* User list */
+.user-list__group {
+  padding-left: 0;
+}
+
+.user-list__item {
+  list-style: none;
+  display: flex;
+  justify-content: space-between;
+  margin-top: 12px;
+  border-bottom: 1px solid #dadada;
+  padding-bottom: 12px;
+  transition: all 0.3s;
+  background-color: #f8fafb;
+  padding: 12px;
+  cursor: pointer;
+  border-radius: 5px;
+}
+
+.user-list__item.not-seen {
+  background-color: #fff;
+  box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
+}
+
+.user-list__item:hover {
+  background-color: #fff;
+  box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
+}
+
+.user-list__item .info p {
+  margin: 0;
+  font-size: 1.1rem;
+  color: var(--primary-color);
+  font-weight: 500;
+}
+
+.user-list__item .info span {
+  font-size: 0.8rem;
+  white-space: nowrap;
+  max-width: 185px;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  display: -webkit-box;
+}
+
+.user-list__item .notification {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: end;
+}
+
+.user-list__item .notification p {
+  margin: 0;
+  font-size: 0.8rem;
+  font-weight: 600;
+  margin-bottom: 4px;
+}
+
+.user-list__item .notification .quantity {
+  font-size: 0.8rem;
+  height: 20px;
+  width: 20px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 50%;
+  background-color: var(--button-color);
+  color: #fff;
+}
