+++ auntion-product/src/app/admin/footer-admin/footer-admin.component.css
