+++ auntion-product/src/app/admin/navbar-admin/navbar-admin.component.html
@@ -1,0 +1,29 @@
+<div class="navbar__admin-container">
+  <div class="navbar__admin-header">
+    <img src="./assets/user/avatar.avif" alt="">
+
+    <span>A0721I1 Admin</span>
+  </div>
+
+  <div class="navbar__admin-list">
+    <div class="navbar__admin-item active">
+      <a [routerLink]="['../']">Dashboard</a>
+    </div>
+
+    <div class="navbar__admin-item">
+      <a [routerLink]="['./chat-page']">Admin Consultant</a>
+    </div>
+
+    <div class="navbar__admin-item">
+      <a>Dashboard</a>
+    </div>
+
+    <div class="navbar__admin-item">
+      <a>Dashboard</a>
+    </div>
+
+    <div class="navbar__admin-item">
+      <a>Dashboard</a>
+    </div>
+  </div>
+</div>
