+++ auntion-product/src/app/admin/statistical/statistical-page/statistical-page.component.html
@@ -1,0 +1,16 @@
+<div class="statistical__admin-container" style="overflow: hidden">
+  <div class="header__admin">
+    <app-header-admin></app-header-admin>
+  </div>
+  <div class="content__admin-container">
+    <div class="row">
+      <div class="d-none-m-b col-md-3 p-r-0">
+        <app-navbar-admin></app-navbar-admin>
+      </div>
+
+      <div class="col-12 col-md-9 p-l-0">
+        <app-footer-admin></app-footer-admin>
+      </div>
+    </div>
+  </div>
+</div>
