+++ auntion-product/src/app/admin/statistical/statistical-page/statistical-page.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-statistical-page',
+  templateUrl: './statistical-page.component.html',
+  styleUrls: ['./statistical-page.component.css']
+})
+export class StatisticalPageComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
