+++ auntion-product/src/app/app-routing.module.ts
@@ -33,8 +33,7 @@ const routes: Routes = [
   },
   {
     path: 'admin',
-    loadChildren: () => import('./control-admin-page/control-admin-page.module').then(module => module.ControlAdminPageModule)
-
+    loadChildren: () => import('./admin/admin.module').then(module => module.AdminModule)
   }
 ];
