+++ auntion-product/src/app/app.component.html
@@ -27,7 +27,7 @@
 <!--<button routerLink="/admin/admin-home">admin-home</button>-->
 
 <!--&lt;!&ndash; bạn nào pull về thì chỉ để lại 4 cái này app-header, router-outlet, app-footer, app-room-chat-user -> còn lại bỏ hết ra nha&ndash;&gt;-->
-<!--<div class="container-app">-->
+<!--<div *ngIf="isUser" class="container-app">-->
 <!--  <app-room-chat-user></app-room-chat-user>-->
 <!--  <app-header class="app-header"></app-header>-->
 <!--  <router-outlet class="router-outlet"></router-outlet>-->
