+++ auntion-product/src/app/app.module.ts
@@ -1,10 +1,9 @@
 import {BrowserModule} from '@angular/platform-browser';
 import {NgModule} from '@angular/core';
-import {HttpClient, HttpClientModule} from '@angular/common/http';
+import {HttpClientModule} from '@angular/common/http';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {AngularFireDatabaseModule} from '@angular/fire/database';
 import {AngularFireAuthModule} from '@angular/fire/auth';
-import {ControlAdminPageModule} from './control-admin-page/control-admin-page.module';
 import {ChatUserPageModule} from './chat-app/chat-user-page/chat-user-page.module';
 import {HomeModule} from './home/home.module';
 import {AppComponent} from './app.component';
@@ -23,7 +22,6 @@ import {AppRoutingModule} from './app-routing.module';
     FormsModule,
     AngularFireDatabaseModule,
     AngularFireAuthModule,
-    ControlAdminPageModule,
     ChatUserPageModule,
     HomeModule,
   ],
