--- auntion-product/src/app/chat-app/chat-admin-page/chat-admin-home/chat-admin-home.component.css
@@ -1,7 +1,0 @@
-.pd-r-0 {
-  padding-right: 0px !important;
-}
-
-.pd-l-0 {
-  padding-left: 0 !important;
-}
