--- auntion-product/src/app/chat-app/chat-admin-page/chat-admin-home/chat-admin-home.component.html
@@ -1,21 +1,0 @@
-<div class="chat-admin__container">
-  <div class="header" style="border-bottom: 1px solid #dadada;"><h3>Header</h3></div>
-  <div class="row">
-    <div style="border-right: 1px solid #dadada" class="col-3">
-<!--      <h3 *ngIf="userItemChild">{{userItemChild}}</h3>-->
-    </div>
-
-    <div class="col-9">
-      <div class="chat-box__container">
-        <div class="row">
-          <div class="col-4" style=" background: #f8fafb;border-right: 1px solid #dadada;">
-            <app-user-list (userIdItem)="getUserFromChild($event)"></app-user-list>
-          </div>
-          <div class="col-8" style="padding: 50px 30px 20px 20px; background: #f8fafb">
-            <app-chat-box [userIdParent]="userItemChild"></app-chat-box>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
