--- auntion-product/src/app/chat-app/chat-admin-page/chat-admin-home/chat-admin-home.component.ts
@@ -1,21 +1,0 @@
-import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
-
-@Component({
-  selector: 'app-chat-admin-home',
-  templateUrl: './chat-admin-home.component.html',
-  styleUrls: ['./chat-admin-home.component.css']
-})
-export class ChatAdminHomeComponent implements OnInit {
-
-  userItemChild: any;
-
-  constructor() { }
-
-  ngOnInit(): void {
-
-  }
-
-  getUserFromChild(event: any) {
-    this.userItemChild = event;
-  }
-}
