+++ auntion-product/src/app/chat-app/chat-user-page/chat-box-user/chat-box-user.component.ts
@@ -1,17 +1,17 @@
 import {Component, OnChanges, OnInit} from '@angular/core';
-import {ApiService} from "../../services/api.service";
-import {Member} from "../../../model/Member";
-import {Account} from "../../../model/Account";
-import {Observable} from "rxjs";
-import {ChatService} from "../../services/chat.service";
-import {Router} from "@angular/router";
+import {ApiService} from '../../services/api.service';
+import {Member} from '../../../model/Member';
+import {Account} from '../../../model/Account';
+import {Observable} from 'rxjs';
+import {ChatService} from '../../services/chat.service';
+import {Router} from '@angular/router';
 
 @Component({
   selector: 'app-chat-box-user',
   templateUrl: './chat-box-user.component.html',
   styleUrls: ['./chat-box-user.component.css']
 })
-export class ChatBoxUserComponent implements OnInit , OnChanges {
+export class ChatBoxUserComponent implements OnInit, OnChanges {
 
   /* Get user */
   member: Member;
@@ -23,8 +23,8 @@ export class ChatBoxUserComponent implements OnInit , OnChanges {
   /* Store message */
   messages: Observable<any>;
 
-  constructor(private apiService: ApiService , private chatService: ChatService
-  ,private router: Router) {
+  constructor(private apiService: ApiService, private chatService: ChatService
+    , private router: Router) {
   }
 
   ngOnInit(): void {
