--- auntion-product/src/app/control-admin-page/content-admin/content-admin.component.html
@@ -1,2 +1,0 @@
-<div class="content-admin" style="height: 100vh; background: #fff">alo
-</div>
