--- auntion-product/src/app/control-admin-page/content-admin/content-admin.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-content-admin',
-  templateUrl: './content-admin.component.html',
-  styleUrls: ['./content-admin.component.css']
-})
-export class ContentAdminComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
