--- auntion-product/src/app/control-admin-page/error500/error500.component.html
@@ -1,18 +1,0 @@
-<div class="container">
-  <div class="row">
-    <div class="col-lg-6 col-lg-offset-3 p404 centered">
-      <img src="../../../assets/img/500.png" alt="">
-      <h1>ERROR 500</h1>
-      <h3>Ouch!! Something Went Wrong!</h3>
-      <br>
-      <div class="row">
-        <div class="col-md-8 col-md-offset-2">
-          <input type="text" class="form-control" id="form1Name" name="form1Name">
-          <button class="btn btn-theme mt">Search</button>
-        </div>
-      </div>
-      <h5 class="mt">Hey, maybe you will be interested in these pages:</h5>
-      <p><a href="/">Index</a> | <a href="#">Sitemap</a> | <a href="/"> Contact</a></p>
-    </div>
-  </div>
-</div>
