--- auntion-product/src/app/control-admin-page/error500/error500.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-error500',
-  templateUrl: './error500.component.html',
-  styleUrls: ['./error500.component.css']
-})
-export class Error500Component implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
