--- auntion-product/src/app/control-admin-page/home-admin/home-admin.component.html
@@ -1,17 +1,0 @@
-<div class="home__admin-container" style="overflow: hidden">
-  <div class="header__admin">
-    <app-header-admin></app-header-admin>
-  </div>
-  <div class="content__admin-container">
-    <div class="row">
-      <div class="d-none-m-b col-md-3 p-r-0">
-        <app-navbar-admin></app-navbar-admin>
-      </div>
-
-      <div class="col-12 col-md-9 p-l-0">
-        <app-content-admin></app-content-admin>
-        <app-footer-admin></app-footer-admin>
-      </div>
-    </div>
-  </div>
-</div>
