--- auntion-product/src/app/control-admin-page/navbar-admin/navbar-admin.component.css
@@ -1,50 +1,0 @@
-.navbar__admin-container {
-  padding: 20px 12px;
-  background-color: #424145;
-  height: 100%;
-}
-
-.navbar__admin-header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.navbar__admin-header img {
-  height: 65px;
-  width: 65px;
-  border-radius: 50%;
-  box-shadow: rgba(0, 0, 0, 0.15) 1.95px 1.95px 2.6px;
-}
-
-.navbar__admin-header span {
-  color: #fff;
-  margin-top: 12px;
-  font-size: 1rem;
-}
-
-/* Navbar list */
-.navbar__admin-list {
-  margin-top: 24px;
-}
-
-.navbar__admin-item {
-  padding: 12px;
-  transition: all 0.3s;
-}
-
-.navbar__admin-item.active {
-  background-color: var(--active-color);
-}
-
-.navbar__admin-item:hover {
-  cursor: pointer;
-  background-color: var(--active-color);
-}
-
-.navbar__admin-item a {
-  display: block;
-  font-size: 0.9rem;
-  color: #fff;
-}
